Refactor FlagMetadata class to inherit from BaseMetadata
