+++ src/OpenFeature/Model/FlagMetadata.cs
@@ -5,53 +5,13 @@ using System.Collections.Immutable;
 #nullable enable
 namespace OpenFeature.Model;
 
-public sealed class FlagMetadata
+public sealed class FlagMetadata : BaseMetadata
 {
-    private readonly ImmutableDictionary<string, object> _metadata;
-
-    public FlagMetadata(Dictionary<string, object> metadata)
-    {
-        this._metadata = metadata.ToImmutableDictionary();
-    }
-
-    public FlagMetadata() : this([])
+    public FlagMetadata()
     {
     }
 
-    public bool? GetBool(string key)
+    public FlagMetadata(Dictionary<string, object> metadata) : base(metadata)
     {
-        return this.GetValue<bool>(key);
-    }
-
-    public int? GetInt(string key)
-    {
-        return this.GetValue<int>(key);
-    }
-
-    public double? GetDouble(string key)
-    {
-        return this.GetValue<double>(key);
-    }
-
-    public string? GetString(string key)
-    {
-        var hasValue = this._metadata.TryGetValue(key, out var value);
-        if (!hasValue)
-        {
-            return null;
-        }
-
-        return value as string ?? throw new InvalidCastException($"Cannot cast {value?.GetType()} to {typeof(string)}");
-    }
-
-    private T? GetValue<T>(string key) where T : struct
-    {
-        var hasValue = this._metadata.TryGetValue(key, out var value);
-        if (!hasValue)
-        {
-            return null;
-        }
-
-        return value is T tValue ? tValue : throw new InvalidCastException($"Cannot cast {value?.GetType()} to {typeof(T)}");
     }
 }
